requirements.txt
amqp==2.6.1
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.2
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.0.4
isort==5.10.1
kombu==4.6.11
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
psycopg2==2.9.3
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
vine==1.3.0
whitenoise==5.3.0
wrapt==1.13.3
